Group tests on unread messages
