+++ python/jupyterlab-collaborative-chat/ui-tests/tests/unread.spec.ts-snapshots/navigation-bottom-linux.png
