+++ python/jupyterlab-collaborative-chat/ui-tests/tests/unread.spec.ts-snapshots/navigation-bottom-unread-linux.png
